[update] rename module system to cmdline
